+++ lcs/agents/racs/components/alp.py
@@ -1,4 +1,5 @@
 from typing import Optional
+import random
 
 from lcs import Perception
 from lcs.agents.racs import Classifier
@@ -7,6 +8,7 @@ from lcs.agents.racs import Classifier
 def expected_case(cl: Classifier,
                   perception: Perception,
                   time: int) -> Optional[Classifier]:
+
     diff = cl.mark.get_differences(perception)
 
     if diff.specificity == 0:
@@ -17,8 +19,30 @@ def expected_case(cl: Classifier,
     no_spec_new = diff.specificity
     child = cl.copy_from(cl, time)
 
-    # TODO: continue from here
-    return None
+    if no_spec >= cl.cfg.u_max:
+        while no_spec >= cl.cfg.u_max:
+            res = cl.generalize_unchanging_condition_attribute()
+            assert res is True
+            no_spec -= 1
+
+        while no_spec + no_spec_new > cl.cfg.u_max:
+            if random.random() < 0.5:
+                diff.generalize_specific_attribute_randomly()
+                no_spec_new -= 1
+            else:
+                if cl.generalize_unchanging_condition_attribute():
+                    no_spec -= 1
+    else:
+        while no_spec + no_spec_new > cl.cfg.u_max:
+            diff.generalize_specific_attribute_randomly()
+            no_spec_new -= 1
+
+    child.condition.specialize(new_condition=diff)
+
+    if child.q < 0.5:
+        child.q = 0.5
+
+    return child
 
 
 def unexpected_case(cl: Classifier,