+++ tests/lcs/agents/racs/test_Condition.py
@@ -36,6 +36,24 @@ class TestCondition:
         # then
         assert cond == Condition(_generalized, cfg)
 
+    @pytest.mark.parametrize("_condition, _spec_before, _spec_after", [
+        ([UBR(2, 6), UBR(7, 2)], 2, 1),
+        ([UBR(2, 6), UBR(0, 16)], 1, 0),
+        ([UBR(0, 16), UBR(0, 16)], 0, 0),
+    ])
+    def test_should_generalize_specific_attributes_randomly(
+            self, _condition, _spec_before, _spec_after, cfg):
+
+        # given
+        condition = Condition(_condition, cfg)
+        assert condition.specificity == _spec_before
+
+        # when
+        condition.generalize_specific_attribute_randomly()
+
+        # then
+        assert condition.specificity == _spec_after
+
     @pytest.mark.parametrize("_condition, _specificity", [
         ([UBR(0, 16), UBR(0, 16)], 0),
         ([UBR(0, 16), UBR(2, 16)], 1),