+++ lcs/agents/racs/Changes.md
@@ -15,3 +15,4 @@
 # Thoughts
 - Maybe effect could return just encoded value, instead of UBR...
 - Specificity/generality should measure how wide is the UBR
+- Maybe u_max should hold information how specific condition should be (not just wildcards but spread)